Radar 그래프 DB 값 토대로 회원평균중량작성,DB userWeight Float타입으로 수정필요

+++ OrderSystem/src/main/java/kr/ac/kopo/model/ExerciseJournal.java
@@ -10,7 +10,7 @@ public class ExerciseJournal {
 	private String start;
 	private String id;
 	private String title;
-	private int userWeight;
+	private float userWeight;
 	private String ExerciseName;
 	private int avgLb;
 	private int sumReps;
@@ -65,12 +65,13 @@ public class ExerciseJournal {
 	}
 
 	
-	public int getUserWeight() {
+
+	public float getUserWeight() {
 		return userWeight;
 	}
 
 
-	public void setUserWeight(int userWeight) {
+	public void setUserWeight(float userWeight) {
 		this.userWeight = userWeight;
 	}
